+++ weed/mq/logstore/log_to_parquet.go
@@ -1,4 +1,4 @@
-package parquet_conv
+package logstore
 
 import (
 	"encoding/binary"
